Main function for blocking records given text data

Function that creates shingles (strings with 2 characters), applies approximate nearest neighbour search using RcppHNSW, RcppAnnoy and mlpack and creates blocks using igraph.

Usage

blocking(
  x,
  y = NULL,
  deduplication = TRUE,
  block = NULL,
  ann = c("hnsw", "lsh", "annoy", "kd"),
  distance = c("cosine", "euclidean", "l2", "ip", "manhatan", "hamming", "angular"),
  ann_read = NULL,
  ann_save = NULL,
  true_blocks = NULL,
  verbose = c(0, 1, 2),
  n_threads = 1,
  seed = 2023,
  control_txt = controls_txt(),
  control_ann = controls_ann()
)

Arguments

x

reference data (character vector or matrix),

y

query data, if not provided NULL by default,

deduplication

whether deduplication should be applied (default TRUE as y is set to NULL),

block

initial blocking to reduce comparisons (currently not supported),

ann

algorithm to be used for searching for ann (possible, c("hnsw", "lsh", "annoy", "kd"), default "hnsw"),

distance

distance metric (default cosine),

ann_read

reading index from file (currently not supported),

ann_save

saving index to file. Two files will be created: 1) with index, 2) with column names (currently not supported),

true_blocks

matrix with true blocks to calculate evaluation metrics (all metrics from igraph::compare() are returned).

verbose

whether log should be provided (0 = none, 1 = main, 2 = ann algorithms),

n_threads

number of threads used for the ann,

seed

seed for the algorithms,

control_txt

list of controls for text data,

control_ann

list of controls for ann algorithms.

Value

Returns a list with containing:

result -- data.frame with indices (rows) of x, y and block

ann -- name of the ann algorithm used,

metrics -- metrics, if true_blocks is provided,

colnames -- variable names (colnames) used for search.

Details

Imports

Author

Maciej Beręsewicz

Examples

## general example
df_example <- data.frame(txt = c("jankowalski", "kowalskijan", "kowalskimjan",
"kowaljan", "montypython", "pythonmonty", "cyrkmontypython", "monty"))

result <- blocking(x = df_example$txt,
                   ann = "hnsw",
                   control_ann = controls_ann(hnsw = list(M = 5, ef_c = 10, ef_s = 10)))
#> 'as(<dgTMatrix>, "dgCMatrix")' is deprecated.
#> Use 'as(., "CsparseMatrix")' instead.
#> See help("Deprecated") and help("Matrix-deprecated").

result
#> Blocking based on the hnsw method.
#> Number of blocks: 2.
#> Number of columns used for blocking: 28.
#> Distribution of the size of the blocks:
#> 1 
#> 2 

## an example with true blocks