Block records based on text data.

Function creates shingles (strings with 2 characters, default) or vectors using a given model (e.g., GloVe), applies approximate nearest neighbour (ANN) algorithms via the rnndescent, RcppHNSW, RcppAnnoy and mlpack packages, and creates blocks using graphs via igraph.

Usage

blocking(
  x,
  y = NULL,
  representation = c("shingles", "vectors"),
  model,
  deduplication = TRUE,
  on = NULL,
  on_blocking = NULL,
  ann = c("nnd", "hnsw", "annoy", "lsh", "kd"),
  distance = c("cosine", "euclidean", "l2", "ip", "manhatan", "hamming", "angular"),
  ann_write = NULL,
  ann_colnames = NULL,
  true_blocks = NULL,
  verbose = c(0, 1, 2),
  graph = FALSE,
  seed = 2023,
  n_threads = 1,
  control_txt = controls_txt(),
  control_ann = controls_ann()
)

Arguments

x

reference data (a character vector or a matrix),

y

query data (a character vector or a matrix), if not provided NULL by default and thus deduplication is performed,

representation

method of representing input data (possible c("shingles", "vectors"); default "shingles"),

model

a matrix containing word embeddings (e.g., GloVe), required only when representation = "vectors",

deduplication

whether deduplication should be applied (default TRUE as y is set to NULL),

on

variables for ANN search (currently not supported),

on_blocking

variables for blocking records before ANN search (currently not supported),

ann

algorithm to be used for searching for ann (possible, c("nnd", "hnsw", "annoy", "lsh", "kd"), default "nnd" which corresponds to nearest neighbour descent method),

distance

distance metric (default cosine, more options are possible see details),

ann_write

writing an index to file. Two files will be created: 1) an index, 2) and text file with column names,

ann_colnames

file with column names if x or y are indices saved on the disk (currently not supported),

true_blocks

matrix with true blocks to calculate evaluation metrics (standard metrics based on confusion matrix as well as all metrics from compare are returned).

verbose

whether log should be provided (0 = none, 1 = main, 2 = ANN algorithm verbose used),

graph

whether a graph should be returned (default FALSE),

seed

seed for the algorithms (for reproducibility),

n_threads

number of threads used for the ANN algorithms and adding data for index and query,

control_txt

list of controls for text data (passed only to itoken_parallel or itoken), used only when representation = "shingles",

control_ann

list of controls for the ANN algorithms.

Value

Returns a list containing:

result – data.table with indices (rows) of x, y, block and distance between points

method – name of the ANN algorithm used,

deduplication – information whether deduplication was applied,

representation – information whether shingles or vectors were used,

metrics – metrics for quality assessment, if true_blocks is provided,

confusion – confusion matrix, if true_blocks is provided,

colnames – variable names (colnames) used for search,

graph – igraph class object.

Author

Maciej Beręsewicz, Adam Struzik

Examples

## an example using RcppHNSW

df_example <- data.frame(txt = c("jankowalski", "kowalskijan", "kowalskimjan",
"kowaljan", "montypython", "pythonmonty", "cyrkmontypython", "monty"))

result <- blocking(x = df_example$txt,
                   ann = "hnsw",
                   control_ann = controls_ann(hnsw = control_hnsw(M = 5, ef_c = 10, ef_s = 10)))

result
#> ========================================================
#> Blocking based on the hnsw method.
#> Number of blocks: 2.
#> Number of columns used for blocking: 28.
#> Reduction ratio: 0.5714.
#> ========================================================
#> Distribution of the size of the blocks:
#> 4 
#> 2 

## an example using mlpack::lsh

result_lsh <- blocking(x = df_example$txt,
                       ann = "lsh")

result_lsh
#> ========================================================
#> Blocking based on the lsh method.
#> Number of blocks: 2.
#> Number of columns used for blocking: 28.
#> Reduction ratio: 0.5714.
#> ========================================================
#> Distribution of the size of the blocks:
#> 4 
#> 2 

## an example using GloVe and RcppAnnoy
if (FALSE) { # \dontrun{
options(timeout = 500)
utils::download.file("https://nlp.stanford.edu/data/glove.6B.zip", destfile = "glove.6B.zip")
utils::unzip("glove.6B.zip")

glove_6B_50d <- readr::read_table("glove.6B.50d.txt",
                                  col_names = FALSE,
                                  show_col_types = FALSE)
data.table::setDT(glove_6B_50d)

glove_vectors <- glove_6B_50d[,-1]
glove_vectors <- as.matrix(glove_vectors)
rownames(glove_vectors) <- glove_6B_50d$X1

## spaces between words are required
df_example_spaces <- data.frame(txt = c("jan kowalski", "kowalski jan", "kowalskim jan",
"kowal jan", "monty python", "python monty", "cyrk monty python", "monty"))

result_annoy <- blocking(x = df_example_spaces$txt,
                         ann = "annoy",
                         representation = "vectors",
                         model = glove_vectors)

result_annoy
} # }

## an example with the NN descent algorithm and true blocks

data(census)
data(cis)

set.seed(2024)
census <- census[sample(nrow(census), floor(nrow(census) / 2)), ]
set.seed(2024)
cis <- cis[sample(nrow(cis), floor(nrow(cis) / 2)), ]

census[, txt:=paste0(pername1, pername2, sex,
       dob_day, dob_mon, dob_year, enumcap, enumpc)]
#>            person_id pername1  pername2    sex dob_day dob_mon dob_year hse_num
#>               <char>   <char>    <char> <char>  <char>  <char>   <char>   <num>
#>     1: SW122AB001001   GEURGE    HUGHES      M      19       5     1942       1
#>     2: SO527ST097003 ISABELLA     MUUBE      F       6       5     2003      97
#>     3:  M820GL006002     EVIE      SHAW      F      29      11     1974       6
#>     4: NW374JN023002    ELLIE      HALL      F      30       1     1950      23
#>     5: NW436BL023001      MIA    HARRIS      F      13       6     1917      23
#>    ---                                                                         
#> 12667: PO774UX023004     ISIA WIIKINSON      F      27       4     1997      23
#> 12668: NW917FN007001    AISHA      HUNT      F      22       2     1965       7
#> 12669: HR559UH032001    YUSUF    MORGAN      M      14      11     1968      32
#> 12670: PO398RR037005    MEGRW   CHRPMRW      F      16       1     2004      37
#> 12671: NW908RT067001     JACK    FOSTER      M       2      10     1933      67
#>                    enumcap  enumpc          str_nam              cap_add
#>                     <char>  <char>           <char>               <char>
#>     1:     1 VICTORIA ROAD SW122AB    Victoria Road     1, Victoria Road
#>     2:        97 MAIN ROAD SO527ST        Main Road        97, Main Road
#>     3:          6 KINGSWAY  M820GL         Kingsway          6, Kingsway
#>     4:     23 STATION ROAD NW374JN     Station Road     23, Station Road
#>     5:      23 FOREST ROAD NW436BL      Forest Road      23, Forest Road
#>    ---                                                                  
#> 12667:      23 LONDON ROAD PO774UX      London Road      23, London Road
#> 12668:        7 GREEN LANE NW917FN       Green Lane        7, Green Lane
#> 12669:     32 STATION ROAD HR559UH     Station Road     32, Station Road
#> 12670:   37 HIGHFIELD ROAD PO398RR   Highfield Road   37, Highfield Road
#> 12671: 67 SPRINGFIELD ROAD NW908RT Springfield Road 67, Springfield Road
#>                census_id                                          txt
#>                   <char>                                       <char>
#>     1: CENSSW122AB001001   GEURGEHUGHESM19519421 VICTORIA ROADSW122AB
#>     2: CENSSO527ST097003      ISABELLAMUUBEF65200397 MAIN ROADSO527ST
#>     3:  CENSM820GL006002            EVIESHAWF291119746 KINGSWAYM820GL
#>     4: CENSNW374JN023002      ELLIEHALLF301195023 STATION ROADNW374JN
#>     5: CENSNW436BL023001       MIAHARRISF136191723 FOREST ROADNW436BL
#>    ---                                                               
#> 12667: CENSPO774UX023004   ISIAWIIKINSONF274199723 LONDON ROADPO774UX
#> 12668: CENSNW917FN007001         AISHAHUNTF22219657 GREEN LANENW917FN
#> 12669: CENSHR559UH032001   YUSUFMORGANM1411196832 STATION ROADHR559UH
#> 12670: CENSPO398RR037005 MEGRWCHRPMRWF161200437 HIGHFIELD ROADPO398RR
#> 12671: CENSNW908RT067001 JACKFOSTERM210193367 SPRINGFIELD ROADNW908RT
cis[, txt:=paste0(pername1, pername2, sex,
    dob_day, dob_mon, dob_year, enumcap, enumpc)]
#>            person_id pername1 pername2    sex dob_day dob_mon dob_year
#>               <char>   <char>   <char> <char>  <char>  <char>   <char>
#>     1: HR802JU071003        I   BUTLER      M      29       5     2004
#>     2: DE921PR018003        C  SIMPSON      F      19       9     2000
#>     3: SW791GH013001     RUBY  ROBERTS      F       3      10     1971
#>     4: PO200HH081002   DARCEY   PARKER   <NA>      19       3     1975
#>     5:  M518DT037003    LEWIS      LEE      M      25      11     1975
#>    ---                                                                
#> 12302: DE921PR017003     ERIN  STEVENS      F      15       4     1969
#> 12303: SW293NE018003    AIDAN    SCOTT      M      12      10     2001
#> 12304: EX264WU014007   SIENNA  JACKSON      F      26       2     1975
#> 12305: HR802JR015002    ABBIE   CARTER      F       3      10     1939
#> 12306: SW791GH008001   SAMUEL    SMITM      M      25       6     1977
#>                    enumcap  enumpc           cis_id
#>                     <char>  <char>           <char>
#>     1:         71 NEW ROAD HR802JU CISHR802JU071003
#>     2:    18 ALBERT STREET DE921PR CISDE921PR018003
#>     3:     13 NORTH STREET SW791GH CISSW791GH013001
#>     4:      81 ALBERT ROAD PO200HH CISPO200HH081002
#>     5:   37 HIGHFIELD ROAD  M518DT  CISM518DT037003
#>    ---                                             
#> 12302:    19 ALBERT STREET DE921PR CISDE921PR017003
#> 12303: 18 MARLBOROUGH ROAD SW293NE CISSW293NE018003
#> 12304: 14 MARLBOROUGH ROAD EX264WU CISEX264WU014007
#> 12305:         15 NEW ROAD HR802JR CISHR802JR015002
#> 12306:      8 NORTH STREET SW791QH CISSW791GH008001
#>                                                    txt
#>                                                 <char>
#>     1:               IBUTLERM295200471 NEW ROADHR802JU
#>     2:         CSIMPSONF199200018 ALBERT STREETDE921PR
#>     3:       RUBYROBERTSF310197113 NORTH STREETSW791GH
#>     4:      DARCEYPARKERNA193197581 ALBERT ROADPO200HH
#>     5:        LEWISLEEM2511197537 HIGHFIELD ROADM518DT
#>    ---                                                
#> 12302:      ERINSTEVENSF154196919 ALBERT STREETDE921PR
#> 12303:   AIDANSCOTTM1210200118 MARLBOROUGH ROADSW293NE
#> 12304: SIENNAJACKSONF262197514 MARLBOROUGH ROADEX264WU
#> 12305:           ABBIECARTERF310193915 NEW ROADHR802JR
#> 12306:        SAMUELSMITMM25619778 NORTH STREETSW791QH

matches <- merge(x = census[, .(x=1:.N, person_id)],
                 y = cis[, .(y = 1:.N, person_id)],
                 by = "person_id")
matches[, block:=1:.N]
#> Key: <person_id>
#>           person_id     x     y block
#>              <char> <int> <int> <int>
#>    1:  DE03US001003  1357  6998     1
#>    2:  DE03US005001  9527  5636     2
#>    3:  DE03US006001 10393 10457     3
#>    4:  DE03US008001  4506  1203     4
#>    5:  DE03US012002  2706  4384     5
#>   ---                                
#> 6077: SW990JE006001 10966  3259  6077
#> 6078: SW990JE006004  7192  4120  6078
#> 6079: SW990JE010001  8217  6327  6079
#> 6080: SW990JE011004  4212 10977  6080
#> 6081: SW990JE012001  2331  9335  6081

result_true_blocks <- blocking(x = census$txt, y = cis$txt, verbose = 1,
                               true_blocks = matches[, .(x, y, block)],
                               seed = 2024)
#> ===== creating tokens =====
#> ===== starting search (nnd, x, y: 12671, 12306, t: 1046) =====
#> ===== creating graph =====

result_true_blocks
#> ========================================================
#> Blocking based on the nnd method.
#> Number of blocks: 8456.
#> Number of columns used for blocking: 1046.
#> Reduction ratio: 0.9999.
#> ========================================================
#> Distribution of the size of the blocks:
#>    2    3    4    5    6    7    8    9   10 
#> 5597 2089  602  131   27    7    1    1    1 
#> ========================================================
#> Evaluation metrics (standard):
#>      recall   precision         fpr         fnr    accuracy specificity 
#>     99.6374     99.4080      0.0001      0.3626     99.9998     99.9999 
#>    f1_score 
#>     99.5226